toolkit/requirements.txt
absl-py==0.13.0
aio-pika==6.8.2
aiofiles==22.1.0
aiohttp==3.7.4
aiormq==3.3.1
APScheduler==3.7.0
arrow==1.2.3
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==21.2.0
bidict==0.22.0
blinker==1.6.2
boto3==1.24.96
botocore==1.27.96
CacheControl==0.12.11
cached-property==1.5.2
cachetools==4.2.4
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
clang==5.0
click==8.1.3
cloudpickle==1.6.0
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.0.4
cryptography==38.0.1
curlify==2.2.1
cycler==0.11.0
dask==2021.11.2
decorator==5.1.1
dm-tree==0.1.7
dnspython==1.16.0
docopt==0.6.2
et-xmlfile==1.1.0
fbmessenger==6.0.0
filelock==3.8.0
fire==0.4.0
Flask==2.3.1
flatbuffers==1.12
fsspec==2022.10.0
future==0.18.2
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.3.post0
grpcio==1.50.0
gunicorn==20.1.0
h5py==3.1.0
httptools==0.5.0
humanfriendly==10.0
idna==3.4
importlib-metadata==5.0.0
iniconfig==1.1.1
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.0.1
jsonpickle==2.0.0
jsonschema==3.2.0
kafka-python==2.0.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
locket==1.0.0
loguru==0.7.0
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.3.4
mattermostwrapper==2.2
MicroHMM==0.4.3
MicroTokenizer==0.21.2
msgpack==1.0.4
multidict==6.0.2
neo4j==4.4.8
networkx==2.6.3
numpy==1.19.5
oauthlib==3.2.2
openai==0.27.6
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==20.9
pamqp==2.3.0
pandas==1.1.5
partd==1.3.0
Pillow==9.2.0
pipreqs==0.4.13
pluggy==1.0.0
prompt-toolkit==2.0.10
protobuf==3.19.6
psycopg2-binary==2.9.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.9
pydot==1.4.2
PyJWT==2.6.0
pykwalify==1.8.0
pymongo==3.10.1
PyMySQL==1.0.3
pyparsing==3.0.9
pyrsistent==0.18.1
pyTelegramBotAPI==3.8.3
pytest==7.1.3
python-crfsuite==0.9.8
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.3.4
python-socketio==5.7.2
pytz==2021.3
PyYAML==6.0
questionary==1.10.0
randomname==0.1.5
rasa==3.0.13
rasa-sdk==3.0.7
redis==3.5.3
regex==2021.8.28
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.10.0
rocketchat-API==1.16.0
rsa==4.9
ruamel.std.argparse==0.8.3
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
sacremoses==0.0.53
sanic==21.12.2
Sanic-Cors==2.2.0
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==0.24.2
scipy==1.7.3
seaborn==0.12.2
sentencepiece==0.1.97
sentry-sdk==1.3.1
setproctitle==1.3.2
six==1.16.0
sklearn-crfsuite==0.3.6
slackclient==2.9.4
SQLAlchemy==1.4.42
tabulate==0.9.0
tarsafe==0.0.3
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.5
tensorflow-addons==0.14.0
tensorflow-estimator==2.6.0
tensorflow-hub==0.12.0
tensorflow-probability==0.13.0
tensorflow-text==2.6.0
termcolor==1.1.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tokenizers==0.7.0
tomli==2.0.1
toolz==0.12.0
tqdm==4.64.1
transformers==2.11.0
twilio==6.50.1
typeguard==2.13.3
typing-utils==0.1.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==2.1
ujson==5.5.0
urllib3==1.26.12
uvloop==0.17.0
wcwidth==0.2.5
webexteamssdk==1.6.1
websockets==10.3
Werkzeug==2.2.2
wrapt==1.12.1
xlrd==1.2.0
yarg==0.1.9
yarl==1.8.1
zipp==3.9.0
